Resolve unbroken ICE subroutines and fix resolution effect execution
Adds resolve_unbroken_subroutines to fire every still-Pending subroutine
on the ICE being encountered, and retargets Effect::BreakSubroutine from
a bare count to a subroutine index so it can drive a specific subroutine.

Also fixes two pre-existing bugs surfaced by review: step_subroutine's
resolve path flipped status to Resolved but never actually applied the
subroutine's effect, and the break/resolve status-transition checks were
duplicated between ability::evaluate_effect and run::engine::step_subroutine.
Both are unified behind a new transition_subroutine helper, which required
changing run::advance_run to mutate GameState in place (so it can call
evaluate_effect) instead of returning a cloned RunState. Also replaces a
raw ICE-index in continue_run's ApproachIce arm with defensive .get().

+++ crates/netrunner_core/src/dsl/effect.rs
@@ -52,8 +52,11 @@ pub enum Effect {
     /// same as `apply_damage` itself.
     DealDamage(DamageType, usize),
     /// Never side-ambiguous — always targets whatever ICE the current run
-    /// is encountering.
-    BreakSubroutine(u32),
+    /// is encountering. Unlike `RunAction::BreakSubroutine`'s index (chosen
+    /// by the player breaking a subroutine manually), this `usize` is the
+    /// index of the specific subroutine this effect itself is wired to
+    /// break — a target, not a count.
+    BreakSubroutine(usize),
     ModifyStrength(i32),
     /// `Side`-explicit for the same reason as `GainCredits`.
     DrawCards(Side, u32),
